requirements.txt
africastalking==1.2.8
asgiref==3.8.1
blinker==1.4
certifi==2024.8.30
charset-normalizer==3.3.2
command-not-found==0.3
coverage==7.6.1
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.1.1
django-allauth==64.2.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
gunicorn==23.0.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-decouple==3.8
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
schema==0.7.7
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.1
systemd-python==234
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
whitenoise==6.7.0
zipp==1.0.0
